Add files via upload
This Java code uses Selenium WebDriver and Apache Commons IO libraries to automate taking a screenshot of the YouTube homepage using the Microsoft Edge browser. The WebDriverManager library is used to manage the driver for the Edge browser. The screenshot is saved in a folder named "Screenshots" with the filename "youtube home page.jpg".